+++ provider/filestore.go
@@ -1,0 +1,215 @@
+package provider
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	awsv2cfg "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/credentials"
+	s3v2 "github.com/aws/aws-sdk-go-v2/service/s3"
+	"gocloud.dev/blob"
+	"gocloud.dev/blob/azureblob"
+	"gocloud.dev/blob/s3blob"
+	"os"
+)
+
+const (
+	Memory     FileStoreType = "MEMORY"
+	FileSystem               = "LOCAL_FILESYSTEM"
+	Azure                    = "AZURE"
+	S3                       = "S3"
+)
+
+type LocalFileStoreConfig struct {
+	DirPath string
+}
+
+func (config *LocalFileStoreConfig) Serialize() ([]byte, error) {
+	data, err := json.Marshal(config)
+	if err != nil {
+		panic(err)
+	}
+	return data, nil
+}
+
+func (config *LocalFileStoreConfig) Deserialize(data []byte) error {
+	err := json.Unmarshal(data, config)
+	if err != nil {
+		return fmt.Errorf("deserialize file blob store config: %w", err)
+	}
+	return nil
+}
+
+type LocalFileStore struct {
+	DirPath string
+	genericFileStore
+}
+
+func NewLocalFileStore(config Config) (FileStore, error) {
+	fileStoreConfig := LocalFileStoreConfig{}
+	if err := fileStoreConfig.Deserialize(config); err != nil {
+		return nil, fmt.Errorf("could not deserialize file store config: %v", err)
+	}
+	bucket, err := blob.OpenBucket(context.TODO(), fileStoreConfig.DirPath)
+	if err != nil {
+		return nil, err
+	}
+	return LocalFileStore{
+		DirPath: fileStoreConfig.DirPath[len("file:///"):],
+		genericFileStore: genericFileStore{
+			bucket: bucket,
+			path:   fileStoreConfig.DirPath,
+		},
+	}, nil
+}
+
+type AzureFileStore struct {
+	AccountName      string
+	AccountKey       string
+	ConnectionString string
+	ContainerName    string
+	Path             string
+	genericFileStore
+}
+
+func (store *AzureFileStore) configString() string {
+	return fmt.Sprintf("fs.azure.account.key.%s.dfs.core.windows.net=%s", store.AccountName, store.AccountKey)
+}
+func (store *AzureFileStore) connectionString() string {
+	return store.ConnectionString
+}
+func (store *AzureFileStore) containerName() string {
+	return store.ContainerName
+}
+
+func (store *AzureFileStore) addAzureVars(envVars map[string]string) map[string]string {
+	envVars["AZURE_CONNECTION_STRING"] = store.ConnectionString
+	envVars["AZURE_CONTAINER_NAME"] = store.ContainerName
+	return envVars
+}
+
+func (store AzureFileStore) AsAzureStore() *AzureFileStore {
+	return &store
+}
+
+type AzureFileStoreConfig struct {
+	AccountName   string
+	AccountKey    string
+	ContainerName string
+	Path          string
+}
+
+func (config *AzureFileStoreConfig) Serialize() ([]byte, error) {
+	data, err := json.Marshal(config)
+	if err != nil {
+		panic(err)
+	}
+	return data, nil
+}
+
+func (config *AzureFileStoreConfig) Deserialize(data []byte) error {
+	err := json.Unmarshal(data, config)
+	if err != nil {
+		return fmt.Errorf("deserialize file blob store config: %w", err)
+	}
+	return nil
+}
+
+func NewAzureFileStore(config Config) (FileStore, error) {
+	azureStoreConfig := AzureFileStoreConfig{}
+	if err := azureStoreConfig.Deserialize(config); err != nil {
+		return nil, fmt.Errorf("could not deserialize azure store config: %v", err)
+	}
+	if err := os.Setenv("AZURE_STORAGE_ACCOUNT", azureStoreConfig.AccountName); err != nil {
+		return nil, fmt.Errorf("could not set storage account env: %w", err)
+	}
+
+	if err := os.Setenv("AZURE_STORAGE_KEY", azureStoreConfig.AccountKey); err != nil {
+		return nil, fmt.Errorf("could not set storage key env: %w", err)
+	}
+	serviceURL := azureblob.ServiceURL(fmt.Sprintf("https://%s.blob.core.windows.net", azureStoreConfig.AccountName))
+	client, err := azureblob.NewDefaultServiceClient(serviceURL)
+	if err != nil {
+		return AzureFileStore{}, fmt.Errorf("could not create azure client: %v", err)
+	}
+
+	bucket, err := azureblob.OpenBucket(context.TODO(), client, azureStoreConfig.ContainerName, nil)
+	if err != nil {
+		return AzureFileStore{}, fmt.Errorf("could not open azure bucket: %v", err)
+	}
+	connectionString := fmt.Sprintf("DefaultEndpointsProtocol=https;AccountName=%s;AccountKey=%s", azureStoreConfig.AccountName, azureStoreConfig.AccountKey)
+	return AzureFileStore{
+		AccountName:      azureStoreConfig.AccountName,
+		AccountKey:       azureStoreConfig.AccountKey,
+		ConnectionString: connectionString,
+		ContainerName:    azureStoreConfig.ContainerName,
+		Path:             azureStoreConfig.Path,
+		genericFileStore: genericFileStore{
+			bucket: bucket,
+		},
+	}, nil
+}
+
+type S3FileStoreConfig struct {
+	AWSAccessKeyId string
+	AWSSecretKey   string
+	BucketRegion   string
+	BucketPath     string
+	Path           string
+}
+
+func (s *S3FileStoreConfig) Deserialize(config SerializedConfig) error {
+	err := json.Unmarshal(config, s)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (s *S3FileStoreConfig) Serialize() []byte {
+	conf, err := json.Marshal(s)
+	if err != nil {
+		panic(err)
+	}
+	return conf
+}
+
+type S3FileStore struct {
+	Bucket string
+	Path   string
+	genericFileStore
+}
+
+func (s *S3FileStore) BlobPath(sourceKey string) string {
+	return sourceKey
+}
+
+func NewS3FileStore(config Config) (FileStore, error) {
+	s3StoreConfig := S3FileStoreConfig{}
+	if err := s3StoreConfig.Deserialize(SerializedConfig(config)); err != nil {
+		return nil, fmt.Errorf("could not deserialize s3 store config: %v", err)
+	}
+	cfg, err := awsv2cfg.LoadDefaultConfig(context.TODO(),
+		awsv2cfg.WithCredentialsProvider(credentials.StaticCredentialsProvider{
+			Value: aws.Credentials{
+				AccessKeyID: s3StoreConfig.AWSAccessKeyId, SecretAccessKey: s3StoreConfig.AWSSecretKey,
+			},
+		}))
+	if err != nil {
+		return nil, err
+	}
+	cfg.Region = s3StoreConfig.BucketRegion
+	clientV2 := s3v2.NewFromConfig(cfg)
+	bucket, err := s3blob.OpenBucketV2(context.TODO(), clientV2, s3StoreConfig.BucketPath, nil)
+	if err != nil {
+		return nil, err
+	}
+	return &S3FileStore{
+		Bucket: s3StoreConfig.BucketPath,
+		Path:   s3StoreConfig.Path,
+		genericFileStore: genericFileStore{
+			bucket: bucket,
+		},
+	}, nil
+}
