+++ provider/k8s.go
@@ -18,7 +18,6 @@ import (
 	"github.com/segmentio/parquet-go"
 	"go.uber.org/zap"
 	"gocloud.dev/blob"
-	"gocloud.dev/blob/azureblob"
 	_ "gocloud.dev/blob/fileblob"
 	_ "gocloud.dev/blob/memblob"
 	"golang.org/x/exp/slices"
@@ -267,13 +266,6 @@ const (
 	K8s                 = "K8S"
 )
 
-const (
-	Memory     FileStoreType = "MEMORY"
-	FileSystem               = "LOCAL_FILESYSTEM"
-	Azure                    = "AZURE"
-	S3                       = "S3"
-)
-
 type K8sConfig struct {
 	ExecutorType   ExecutorType
 	ExecutorConfig interface{}
@@ -468,35 +460,6 @@ type FileStore interface {
 
 type Iterator interface {
 	Next() (map[string]interface{}, error)
-}
-
-type AzureFileStore struct {
-	AccountName      string
-	AccountKey       string
-	ConnectionString string
-	ContainerName    string
-	Path             string
-	genericFileStore
-}
-
-func (store *AzureFileStore) configString() string {
-	return fmt.Sprintf("fs.azure.account.key.%s.dfs.core.windows.net=%s", store.AccountName, store.AccountKey)
-}
-func (store *AzureFileStore) connectionString() string {
-	return store.ConnectionString
-}
-func (store *AzureFileStore) containerName() string {
-	return store.ContainerName
-}
-
-func (store *AzureFileStore) addAzureVars(envVars map[string]string) map[string]string {
-	envVars["AZURE_CONNECTION_STRING"] = store.ConnectionString
-	envVars["AZURE_CONTAINER_NAME"] = store.ContainerName
-	return envVars
-}
-
-func (store AzureFileStore) AsAzureStore() *AzureFileStore {
-	return &store
 }
 
 type genericFileStore struct {
@@ -816,107 +779,6 @@ func (store genericFileStore) Delete(key string) error {
 
 func (store genericFileStore) Close() error {
 	return store.bucket.Close()
-}
-
-type LocalFileStoreConfig struct {
-	DirPath string
-}
-
-func (config *LocalFileStoreConfig) Serialize() ([]byte, error) {
-	data, err := json.Marshal(config)
-	if err != nil {
-		panic(err)
-	}
-	return data, nil
-}
-
-func (config *LocalFileStoreConfig) Deserialize(data []byte) error {
-	err := json.Unmarshal(data, config)
-	if err != nil {
-		return fmt.Errorf("deserialize file blob store config: %w", err)
-	}
-	return nil
-}
-
-type LocalFileStore struct {
-	DirPath string
-	genericFileStore
-}
-
-func NewLocalFileStore(config Config) (FileStore, error) {
-	fileStoreConfig := LocalFileStoreConfig{}
-	if err := fileStoreConfig.Deserialize(config); err != nil {
-		return nil, fmt.Errorf("could not deserialize file store config: %v", err)
-	}
-	bucket, err := blob.OpenBucket(context.TODO(), fileStoreConfig.DirPath)
-	if err != nil {
-		return nil, err
-	}
-	return LocalFileStore{
-		DirPath: fileStoreConfig.DirPath[len("file:///"):],
-		genericFileStore: genericFileStore{
-			bucket: bucket,
-			path:   fileStoreConfig.DirPath,
-		},
-	}, nil
-}
-
-type AzureFileStoreConfig struct {
-	AccountName   string
-	AccountKey    string
-	ContainerName string
-	Path          string
-}
-
-func (config *AzureFileStoreConfig) Serialize() ([]byte, error) {
-	data, err := json.Marshal(config)
-	if err != nil {
-		panic(err)
-	}
-	return data, nil
-}
-
-func (config *AzureFileStoreConfig) Deserialize(data []byte) error {
-	err := json.Unmarshal(data, config)
-	if err != nil {
-		return fmt.Errorf("deserialize file blob store config: %w", err)
-	}
-	return nil
-}
-
-func NewAzureFileStore(config Config) (FileStore, error) {
-	azureStoreConfig := AzureFileStoreConfig{}
-	if err := azureStoreConfig.Deserialize(config); err != nil {
-		return nil, fmt.Errorf("could not deserialize azure store config: %v", err)
-	}
-	if err := os.Setenv("AZURE_STORAGE_ACCOUNT", azureStoreConfig.AccountName); err != nil {
-		return nil, fmt.Errorf("could not set storage account env: %w", err)
-	}
-
-	if err := os.Setenv("AZURE_STORAGE_KEY", azureStoreConfig.AccountKey); err != nil {
-		return nil, fmt.Errorf("could not set storage key env: %w", err)
-	}
-	serviceURL := azureblob.ServiceURL(fmt.Sprintf("https://%s.blob.core.windows.net", azureStoreConfig.AccountName))
-	client, err := azureblob.NewDefaultServiceClient(serviceURL)
-	if err != nil {
-		return AzureFileStore{}, fmt.Errorf("could not create azure client: %v", err)
-	}
-
-	bucket, err := azureblob.OpenBucket(context.TODO(), client, azureStoreConfig.ContainerName, nil)
-	if err != nil {
-		return AzureFileStore{}, fmt.Errorf("could not open azure bucket: %v", err)
-	}
-	connectionString := fmt.Sprintf("DefaultEndpointsProtocol=https;AccountName=%s;AccountKey=%s", azureStoreConfig.AccountName, azureStoreConfig.AccountKey)
-	return AzureFileStore{
-		AccountName:      azureStoreConfig.AccountName,
-		AccountKey:       azureStoreConfig.AccountKey,
-		ConnectionString: connectionString,
-		ContainerName:    azureStoreConfig.ContainerName,
-		Path:             azureStoreConfig.Path,
-		genericFileStore: genericFileStore{
-			bucket: bucket,
-		},
-	}, nil
 }
 
 func ResourcePrefix(id ResourceID) string {
@@ -1579,10 +1441,6 @@ func (k8s *K8sOfflineStore) registeredResourceSchema(id ResourceID) (ResourceSch
 	}
 	k8s.logger.Debugw("Successfully retrieved resource schema", "id", id)
 	return blobResourceTable.schema, nil
-}
-
-func (s *S3FileStore) BlobPath(sourceKey string) string {
-	return sourceKey
 }
 
 func (k8s *K8sOfflineStore) trainingSet(def TrainingSetDef, isUpdate bool) error {
