+++ provider/spark.go
@@ -12,16 +12,11 @@ import (
 	"github.com/featureform/helpers"
 	"github.com/featureform/logging"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	awsv2cfg "github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/credentials"
-	"github.com/aws/aws-sdk-go-v2/service/emr"
-	s3v2 "github.com/aws/aws-sdk-go-v2/service/s3"
-	"go.uber.org/zap"
-	"gocloud.dev/blob/s3blob"
-
 	databricks "github.com/Azure/databricks-sdk-golang"
 	dbAzure "github.com/Azure/databricks-sdk-golang/azure"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/emr"
+	"go.uber.org/zap"
 
 	// clusterHTTPModels "github.com/Azure/databricks-sdk-golang/azure/clusters/httpmodels"
 	// clusterModels "github.com/Azure/databricks-sdk-golang/azure/clusters/models"
@@ -280,65 +275,6 @@ func (db *DatabricksExecutor) RunSparkJob(args *[]string, store FileStore) error
 		time.Sleep(1 * time.Second)
 	}
 	return nil
-}
-
-type S3FileStoreConfig struct {
-	AWSAccessKeyId string
-	AWSSecretKey   string
-	BucketRegion   string
-	BucketPath     string
-	Path           string
-}
-
-func (s *S3FileStoreConfig) Deserialize(config SerializedConfig) error {
-	err := json.Unmarshal(config, s)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *S3FileStoreConfig) Serialize() []byte {
-	conf, err := json.Marshal(s)
-	if err != nil {
-		panic(err)
-	}
-	return conf
-}
-
-type S3FileStore struct {
-	Bucket string
-	Path   string
-	genericFileStore
-}
-
-func NewS3FileStore(config Config) (FileStore, error) {
-	s3StoreConfig := S3FileStoreConfig{}
-	if err := s3StoreConfig.Deserialize(SerializedConfig(config)); err != nil {
-		return nil, fmt.Errorf("could not deserialize s3 store config: %v", err)
-	}
-	cfg, err := awsv2cfg.LoadDefaultConfig(context.TODO(),
-		awsv2cfg.WithCredentialsProvider(credentials.StaticCredentialsProvider{
-			Value: aws.Credentials{
-				AccessKeyID: s3StoreConfig.AWSAccessKeyId, SecretAccessKey: s3StoreConfig.AWSSecretKey,
-			},
-		}))
-	if err != nil {
-		return nil, err
-	}
-	cfg.Region = s3StoreConfig.BucketRegion
-	clientV2 := s3v2.NewFromConfig(cfg)
-	bucket, err := s3blob.OpenBucketV2(context.TODO(), clientV2, s3StoreConfig.BucketPath, nil)
-	if err != nil {
-		return nil, err
-	}
-	return &S3FileStore{
-		Bucket: s3StoreConfig.BucketPath,
-		Path:   s3StoreConfig.Path,
-		genericFileStore: genericFileStore{
-			bucket: bucket,
-		},
-	}, nil
 }
 
 type PythonOfflineQueries interface {
